Introduce "ModifyEndpoints" in extended EndpointModifyingProvider interface
This avoids that providers can only modify endpoints in their
ApplyChanges method, possibly leading to no-op upsert operations because
external-dns can't foresee these automatic changes.

The prime example for this is automatic conversion of CNAME records to
alias records in the Route 53 provider if their target is an ELB/ALB.
When retrieving these records from the zone, the provider adds the
provider-specific properties for "alias" and "evaluate-target-health"
but when extracting the corresponding endpoints from the cluster state,
external-dns will not know that the provider will set these attributes
automatically and will therefor issue a (no-op) update operation.

This change calls a "ModifyEndpoints" method in the provider to let it
do these automatic modifications before external-dns calculates the
change plan, allowing it do skip the no-op update. As a consequence, the
Route 53 provider can move the automatic CNAME to alias conversion in
this method (where it only needs to the add the correct
provider-specific properties) and in ApplyChanges only needs to act upon
these. This also gets rid of the special-case handling of
evaluate-target-health in the plan calculation.